+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- fix: Turn off cssnano's `cssDeclarationSorter` to aid minification of mixins
 - feat: Allow opting out of minification
 
 ## 0.2.1